change i32_shr_u operator data type

+++ src/lib/core/runtime/i32_shr_u.c
@@ -11,7 +11,7 @@ int runtime_i32_shr_u(Stack* stack)
     stack->entries->pop(stack->entries, (void **)&value2);
     stack->entries->pop(stack->entries, (void **)&value1);
 
-    stack->entries->push(stack->entries, new_i32Value((uint32_t)value1->value.i32 >> (uint32_t)value2->value.i32));
+    stack->entries->push(stack->entries, new_i32Value(value1->value.u32 >> value2->value.u32));
 
     free(value1);
     free(value2);
